Expanded model to include Scripts with Parameters
This is really the main thing. Everything else about this web
application is about managing these scripts.  A script contains not only
the file-name for the code that will be run, but a comprehensive list of
the command line parameters that the code might use.  Once completed,
the Script object will be able to render an input form, a list of
helpful comments, or the actual command line to be run, all based on its
parameters.